@@ -1,9 +1,8 @@
 package org.rembx.jeeshop.catalog;
 
-
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.rembx.jeeshop.catalog.model.CatalogItem;
 import org.rembx.jeeshop.catalog.model.CatalogPersistenceUnit;
 import org.rembx.jeeshop.catalog.model.Presentation;
@@ -24,14 +23,14 @@ public class PresentationResourceCT {
 
     private TestCatalog testCatalog;
     private static EntityManagerFactory entityManagerFactory;
-    private EntityManager entityManager;
+    EntityManager entityManager;
 
-    @BeforeClass
+    @BeforeAll
     public static void beforeClass() {
         entityManagerFactory = Persistence.createEntityManagerFactory(CatalogPersistenceUnit.NAME);
     }
 
-    @Before
+    @BeforeEach
     public void setup() {
         testCatalog = TestCatalog.getInstance();
         entityManager = entityManagerFactory.createEntityManager();
@@ -40,7 +39,7 @@ public class PresentationResourceCT {
     @Test
     public void find_shouldReturnPresentation() {
         Presentation presentation = new Presentation("en_GB", "presentation1", "short description", "long description");
-        service = new PresentationResource(null, null, null, presentation);
+        service = new PresentationResource(entityManager, null).init(null, null, presentation);
 
         assertThat(service.find()).isEqualTo(presentation);
     }
@@ -50,7 +49,7 @@ public class PresentationResourceCT {
         CatalogItem parentCatalogItem = testCatalog.aCategoryWithoutPresentation();
 
         Presentation presentation = new Presentation(null, "presentation test", "testShortDesc", "testLongDesc");
-        service = new PresentationResource(entityManager, parentCatalogItem, "fr_FR", null);
+        service = new PresentationResource(entityManager, null).init(parentCatalogItem, "fr_FR", null);
 
         entityManager.getTransaction().begin(); // wrap call in transaction as method is transactional
         service.createLocalizedPresentation(presentation);
@@ -70,7 +69,7 @@ public class PresentationResourceCT {
 
         Presentation presentation = createTestPresentation();
 
-        service = new PresentationResource(entityManager, null, "fr_FR", presentation);
+        service = new PresentationResource(entityManager, null).init(null, "fr_FR", presentation);
 
         UUID uuid = UUID.randomUUID();
         presentation.setShortDescription(uuid.toString());
@@ -93,7 +92,7 @@ public class PresentationResourceCT {
         Presentation presentation = createTestPresentation();
         parentCatalogItem.getPresentationByLocale().put("fr_FR",presentation);
 
-        service = new PresentationResource(entityManager, parentCatalogItem, "fr_FR", presentation);
+        service = new PresentationResource(entityManager, null).init(parentCatalogItem, "fr_FR", presentation);
 
         entityManager.getTransaction().begin(); // wrap call in transaction as method is transactional
         service.delete();
