@@ -1,8 +1,10 @@
 package org.rembx.jeeshop.order;
 
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import io.quarkus.undertow.runtime.HttpSessionContext;
+import org.apache.http.auth.BasicUserPrincipal;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.rembx.jeeshop.catalog.model.CatalogPersistenceUnit;
 import org.rembx.jeeshop.catalog.test.TestCatalog;
 import org.rembx.jeeshop.mail.Mailer;
@@ -15,22 +17,22 @@ import org.rembx.jeeshop.role.JeeshopRoles;
 import org.rembx.jeeshop.user.MailTemplateFinder;
 import org.rembx.jeeshop.user.UserFinder;
 import org.rembx.jeeshop.user.model.Address;
+
 import org.rembx.jeeshop.user.model.User;
 import org.rembx.jeeshop.user.model.UserPersistenceUnit;
-import sun.security.acl.PrincipalImpl;
 
-import javax.ejb.SessionContext;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.fail;
+import static org.assertj.core.api.Assertions.fail;
 import static org.mockito.Mockito.*;
 import static org.rembx.jeeshop.order.model.OrderStatus.PAYMENT_VALIDATED;
 
@@ -38,18 +40,18 @@ public class OrdersCT {
 
     private static EntityManagerFactory emf;
     private static EntityManagerFactory catalogEmf;
-    private EntityManager entityManager;
+    EntityManager entityManager;
     private EntityManager catalogEntityManager;
     private TestOrder testOrder;
     private TestCatalog testCatalog;
 
 
-    private SessionContext sessionContextMock;
+    private SecurityContext sessionContextMock;
     private PriceEngine priceEngineMock;
     private PaymentTransactionEngine paymentTransactionEngine;
     private Orders service;
 
-    @BeforeClass
+    @BeforeAll
     public static void beforeClass() {
         emf = Persistence.createEntityManagerFactory(UserPersistenceUnit.NAME);
         catalogEmf = Persistence.createEntityManagerFactory(CatalogPersistenceUnit.NAME);
@@ -57,32 +59,32 @@ public class OrdersCT {
 
     }
 
-    @Before
+    @BeforeEach
     public void setup() {
         testOrder = TestOrder.getInstance();
         testCatalog = TestCatalog.getInstance();
 
         entityManager = emf.createEntityManager();
         catalogEntityManager = catalogEmf.createEntityManager();
-        sessionContextMock = mock(SessionContext.class);
+        sessionContextMock = mock(SecurityContext.class);
         priceEngineMock = mock(PriceEngine.class);
-        final MailTemplateFinder mailTemplateFinder = new MailTemplateFinder(entityManager);
+        final MailTemplateFinder mailTemplateFinder = MailTemplateFinder.getInstance(entityManager);
         paymentTransactionEngine = new DefaultPaymentTransactionEngine(
                 new OrderFinder(entityManager, catalogEntityManager, new OrderConfiguration("20", "3")),
                 mailTemplateFinder, new Mailer(), catalogEntityManager);
 
         service = new Orders(entityManager, new OrderFinder(entityManager, catalogEntityManager, new OrderConfiguration("11.0", "20.0")), new UserFinder(entityManager),
-                mailTemplateFinder, null, sessionContextMock, priceEngineMock, paymentTransactionEngine);
+                mailTemplateFinder, null, priceEngineMock, paymentTransactionEngine);
     }
 
     @Test
     public void find() throws Exception {
-        assertThat(service.find(1L, null)).isEqualTo(testOrder.firstOrder());
+        assertThat(service.find(sessionContextMock,1L, null)).isEqualTo(testOrder.firstOrder());
     }
 
     @Test
     public void find_withEnhanceResult_shouldReturnEnhancedOrder() throws Exception {
-        Order enhancedOrder = service.find(1L, true);
+        Order enhancedOrder = service.find(sessionContextMock,1L, true);
 
         assertThat(enhancedOrder).isEqualTo(testOrder.firstOrder());
 
@@ -92,7 +94,7 @@ public class OrdersCT {
     @Test
     public void find_withUnknownId_ShouldThrowException() throws Exception {
         try {
-            service.find(999L, null);
+            service.find(sessionContextMock, 999L, null);
             fail("should have thrown ex");
         } catch (WebApplicationException e) {
             assertThat(e.getResponse().getStatusInfo()).isEqualTo(Response.Status.NOT_FOUND);
@@ -106,11 +108,11 @@ public class OrdersCT {
         entityManager.persist(user);
         entityManager.getTransaction().commit();
 
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(true);
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.ADMIN)).thenReturn(false);
-        when(sessionContextMock.getCallerPrincipal()).thenReturn(new PrincipalImpl("777@test.com"));
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(true);
+        when(sessionContextMock.isUserInRole(JeeshopRoles.ADMIN)).thenReturn(false);
+        when(sessionContextMock.getUserPrincipal()).thenReturn(new BasicUserPrincipal("777@test.com"));
         try {
-            service.find(1L, null);
+            service.find(sessionContextMock, 1L, null);
             fail("should have thrown ex");
         } catch (WebApplicationException e) {
             assertThat(e.getResponse().getStatusInfo()).isEqualTo(Response.Status.UNAUTHORIZED);
@@ -124,45 +126,45 @@ public class OrdersCT {
 
     @Test
     public void findAll_shouldReturnNoneEmptyList() {
-        assertThat(service.findAll(null, null, null, null, null, null, null, null)).containsExactly(testOrder.firstOrder(), testOrder.secondOrder());
+        assertThat(service.findAll(sessionContextMock, null, null, null, null, null, null, null, null)).containsExactly(testOrder.firstOrder(), testOrder.secondOrder());
     }
 
     @Test
     public void findAll_withPagination_shouldReturnNoneEmptyListPaginated() {
-        List<Order> orders = service.findAll(null, 0, 1, null, null, null, null, null);
+        List<Order> orders = service.findAll(sessionContextMock, null, 0, 1, null, null, null, null, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
     }
 
     @Test
     public void findAll_ByLogin_shouldReturnSearchedOrder() {
-        List<Order> orders = service.findAll(testOrder.firstOrder().getUser().getLogin(), 0, 1, null, null, null, null, null);
+        List<Order> orders = service.findAll(sessionContextMock, testOrder.firstOrder().getUser().getLogin(), 0, 1, null, null, null, null, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
     }
 
     @Test
     public void findAll_ByStatus_shouldReturnOrdersWithMatchingStatus() {
-        List<Order> orders = service.findAll(null, 0, 1, null, null, OrderStatus.PAYMENT_VALIDATED, null, null);
+        List<Order> orders = service.findAll(sessionContextMock, null, 0, 1, null, null, OrderStatus.PAYMENT_VALIDATED, null, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
     }
 
     @Test
     public void findAll_ByStatus_shouldReturnEmptyListWhenThereAreNoOrdersInGivenStatus() {
-        List<Order> orders = service.findAll(null, 0, 1, null, null, OrderStatus.DELIVERED, null, null);
+        List<Order> orders = service.findAll(sessionContextMock, null, 0, 1, null, null, OrderStatus.DELIVERED, null, null);
         assertThat(orders).isEmpty();
     }
 
     @Test
     public void findAll_ByStatus_shouldReturnEmptyListWhenThereAreNoOrdersWithItemsMatchingGivenSKUId() {
-        List<Order> orders = service.findAll(null, 0, 1, null, null, null, 2L, null);
+        List<Order> orders = service.findAll(sessionContextMock,null, 0, 1, null, null, null, 2L, null);
         assertThat(orders).isEmpty();
     }
 
     @Test
     public void findAll_withEnhancedResults_shouldReturnEnhancedOrders() {
-        List<Order> orders = service.findAll(null, 0, 1, null, null, null, null, true);
+        List<Order> orders = service.findAll(sessionContextMock, null, 0, 1, null, null, null, null, true);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
         assertThat(orders.get(0).getDeliveryFee()).isNotNull();
@@ -172,70 +174,69 @@ public class OrdersCT {
 
     @Test
     public void findAll_BySkuId_shouldReturnOrdersWithItemsHavingGivenSkuIdAndStatus() {
-        List<Order> orders = service.findAll(null, 0, 1, null, null, OrderStatus.PAYMENT_VALIDATED, 1L, null);
+        List<Order> orders = service.findAll(sessionContextMock,null, 0, 1, null, null, OrderStatus.PAYMENT_VALIDATED, 1L, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
     }
 
     @Test
     public void findAll_BySkuId_shouldReturnOrdersWithItemsHavingGivenSkuId() {
-        List<Order> orders = service.findAll(null, 0, 1, null, null, null, 1L, null);
+        List<Order> orders = service.findAll(sessionContextMock,null, 0, 1, null, null, null, 1L, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
     }
 
-
     @Test
     public void findAll_whenClientHasUserRoleOnly_shouldReturnNoneEmptyList_WithoutOrdersWithStatusCREATED() {
 
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(true);
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.ADMIN)).thenReturn(false);
-        when(sessionContextMock.getCallerPrincipal()).thenReturn(new PrincipalImpl(testOrder.firstOrdersUser().getLogin()));
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(true);
+        when(sessionContextMock.isUserInRole(JeeshopRoles.ADMIN)).thenReturn(false);
+        when(sessionContextMock.getUserPrincipal()).thenReturn(new BasicUserPrincipal(testOrder.firstOrdersUser().getLogin()));
 
-        assertThat(service.findAll(null, null, null, null, null, null, null, null)).containsExactly(testOrder.firstOrder());
+        assertThat(service.findAll(sessionContextMock,null, null, null, null, null, null, null, null)).containsExactly(testOrder.firstOrder());
     }
 
     @Test
     public void findAll_whenClientHasUserRoleOnlyAndwithPagination_shouldReturnNoneEmptyListPaginated() {
 
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(true);
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.ADMIN)).thenReturn(false);
-        when(sessionContextMock.getCallerPrincipal()).thenReturn(new PrincipalImpl(testOrder.firstOrdersUser().getLogin()));
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(true);
+        when(sessionContextMock.isUserInRole(JeeshopRoles.ADMIN)).thenReturn(false);
+        when(sessionContextMock.getUserPrincipal()).thenReturn(new BasicUserPrincipal(testOrder.firstOrdersUser().getLogin()));
 
-        List<Order> orders = service.findAll(null, 0, 1, null, null, null, null, null);
+        List<Order> orders = service.findAll(sessionContextMock, null, 0, 1, null, null, null, null, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
     }
 
     @Test
     public void findAll_whenClientHasUserRoleOnlyAndByLogin_shouldReturnSearchedOrder() {
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(true);
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.ADMIN)).thenReturn(false);
-        when(sessionContextMock.getCallerPrincipal()).thenReturn(new PrincipalImpl(testOrder.firstOrdersUser().getLogin()));
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(true);
+        when(sessionContextMock.isUserInRole(JeeshopRoles.ADMIN)).thenReturn(false);
+        when(sessionContextMock.getUserPrincipal()).thenReturn(new BasicUserPrincipal(testOrder.firstOrdersUser().getLogin()));
 
-        List<Order> orders = service.findAll(testOrder.firstOrder().getUser().getLogin(), 0, 1, null, null, null, null, null);
+        List<Order> orders = service.findAll(sessionContextMock, testOrder.firstOrder().getUser().getLogin(), 0, 1, null, null, null, null, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
     }
 
     @Test
     public void findAll_whenClientHasUserRoleOnlyAndByStatus_shouldReturnOrdersWithMatchingStatus() {
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(true);
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.ADMIN)).thenReturn(false);
-        when(sessionContextMock.getCallerPrincipal()).thenReturn(new PrincipalImpl(testOrder.firstOrdersUser().getLogin()));
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(true);
+        when(sessionContextMock.isUserInRole(JeeshopRoles.ADMIN)).thenReturn(false);
+        when(sessionContextMock.getUserPrincipal()).thenReturn(new BasicUserPrincipal(testOrder.firstOrdersUser().getLogin()));
 
-        List<Order> orders = service.findAll(null, 0, 1, null, null, OrderStatus.PAYMENT_VALIDATED, null, null);
+        List<Order> orders = service.findAll(sessionContextMock, null, 0, 1, null, null, OrderStatus.PAYMENT_VALIDATED, null, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.firstOrder());
     }
 
     @Test
     public void findAll_whenClientHasUserRoleAndByStatusCREATED_shouldReturnOrdersWithThisStatus() {
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(true);
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.ADMIN)).thenReturn(false);
-        when(sessionContextMock.getCallerPrincipal()).thenReturn(new PrincipalImpl(testOrder.firstOrdersUser().getLogin()));
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(true);
+        when(sessionContextMock.isUserInRole(JeeshopRoles.ADMIN)).thenReturn(false);
+        when(sessionContextMock.getUserPrincipal()).thenReturn(new BasicUserPrincipal(testOrder.firstOrdersUser().getLogin()));
 
-        List<Order> orders = service.findAll(null, 0, 1, null, null, OrderStatus.CREATED, null, null);
+        List<Order> orders = service.findAll(sessionContextMock, null, 0, 1, null, null, OrderStatus.CREATED, null, null);
         assertThat(orders).isNotEmpty();
         assertThat(orders).containsExactly(testOrder.secondOrder());
     }
@@ -261,7 +262,7 @@ public class OrdersCT {
 
         try {
             Order order = new Order(null, address, new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA"));
-            service.create(order, null);
+            service.create(sessionContextMock, order, null);
             fail("should have thrown ex");
         } catch (WebApplicationException e) {
             assertThat(e.getResponse().getStatusInfo()).isEqualTo(Response.Status.BAD_REQUEST);
@@ -269,7 +270,7 @@ public class OrdersCT {
 
         try {
             Order order = new Order(null, new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA"), address);
-            service.create(order, null);
+            service.create(sessionContextMock, order, null);
             fail("should have thrown ex");
         } catch (WebApplicationException e) {
             assertThat(e.getResponse().getStatusInfo()).isEqualTo(Response.Status.BAD_REQUEST);
@@ -277,7 +278,7 @@ public class OrdersCT {
 
         try {
             Order order = new Order(orderItems, new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA"), address);
-            service.create(order, null);
+            service.create(sessionContextMock, order, null);
             fail("should have thrown ex");
         } catch (WebApplicationException e) {
             assertThat(e.getResponse().getStatusInfo()).isEqualTo(Response.Status.BAD_REQUEST);
@@ -290,18 +291,18 @@ public class OrdersCT {
         Address deliveryAddress = new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA");
         Address billingAddress = new Address("8 Rue Toto", "Paris", "75001", "John", "Doe", "M.", null, "FRA");
 
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(true);
-        when(sessionContextMock.getCallerPrincipal()).thenReturn(new PrincipalImpl(testOrder.firstOrdersUser().getLogin()));
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(true);
+        when(sessionContextMock.getUserPrincipal()).thenReturn(new BasicUserPrincipal(testOrder.firstOrdersUser().getLogin()));
 
         entityManager.getTransaction().begin();
         Order order = new Order(new HashSet<>(), new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA"), new Address("8 Rue Toto", "Paris", "75001", "John", "Doe", "M.", null, "FRA"));
 
         order.setOrderDiscounts(new HashSet<>());
 
-        service.create(order, null);
+        service.create(sessionContextMock, order, null);
         entityManager.getTransaction().commit();
 
-        verify(sessionContextMock).isCallerInRole(JeeshopRoles.USER);
+        verify(sessionContextMock).isUserInRole(JeeshopRoles.USER);
         //verify(mailerMock).sendMail(testMailTemplate.orderConfirmationMailTemplate().getSubject(), testOrder.firstOrdersUser().getLogin(), "<html><body>Hello M. John Doe. Your order has been registered...</body></html>");
 
         final Order persistedOrder = entityManager.find(Order.class, order.getId());
@@ -330,18 +331,18 @@ public class OrdersCT {
         orderItems.add(new OrderItem(2L, 2L, 3));
 
 
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(true);
-        when(sessionContextMock.getCallerPrincipal()).thenReturn(new PrincipalImpl(testOrder.firstOrdersUser().getLogin()));
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(true);
+        when(sessionContextMock.getUserPrincipal()).thenReturn(new BasicUserPrincipal(testOrder.firstOrdersUser().getLogin()));
 
         entityManager.getTransaction().begin();
         Order order = new Order(orderItems, new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA"), new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA"));
         order.setOrderDiscounts(new HashSet<>());
 
-        service.create(order, null);
+        service.create(sessionContextMock, order, null);
         entityManager.getTransaction().commit();
 
 
-        verify(sessionContextMock).isCallerInRole(JeeshopRoles.USER);
+        verify(sessionContextMock).isUserInRole(JeeshopRoles.USER);
         verify(priceEngineMock).computePrice(order);
 
         final Order persistedOrder = entityManager.find(Order.class, order.getId());
@@ -362,18 +363,18 @@ public class OrdersCT {
     @Test
     public void create_shouldSetGivenUserByLoginInOrder_ForADMINRole() throws Exception {
 
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.USER)).thenReturn(false);
-        when(sessionContextMock.isCallerInRole(JeeshopRoles.ADMIN)).thenReturn(true);
+        when(sessionContextMock.isUserInRole(JeeshopRoles.USER)).thenReturn(false);
+        when(sessionContextMock.isUserInRole(JeeshopRoles.ADMIN)).thenReturn(true);
 
         entityManager.getTransaction().begin();
         Order order = new Order(new HashSet<>(), new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA"), new Address("7 Rue des arbres", "Paris", "92800", "John", "Doe", "M.", null, "USA"));
         order.setOrderDiscounts(new HashSet<>());
 
-        service.create(order, "test@test.com");
+        service.create(sessionContextMock, order, "test@test.com");
         entityManager.getTransaction().commit();
 
-        verify(sessionContextMock).isCallerInRole(JeeshopRoles.USER);
-        verify(sessionContextMock).isCallerInRole(JeeshopRoles.ADMIN);
+        verify(sessionContextMock).isUserInRole(JeeshopRoles.USER);
+        verify(sessionContextMock).isUserInRole(JeeshopRoles.ADMIN);
 
         final Order persistedOrder = entityManager.find(Order.class, order.getId());
 
