@@ -12,16 +12,12 @@ import javax.ws.rs.core.Application;
 import java.util.HashSet;
 import java.util.Set;
 
-@ApplicationPath("/rs")
-public class
-        ApplicationConfig extends Application {
-
-    // ======================================
-    // =          Business methods          =
-    // ======================================
+@ApplicationPath("/")
+public class ApplicationConfig extends Application {
 
     @Override
     public Set<Class<?>> getClasses() {
+
         Set<Class<?>> classes = new HashSet<>();
         classes.add(Catalogs.class);
         classes.add(Categories.class);
