@@ -1,11 +1,11 @@
 package org.rembx.jeeshop.catalog.model;
 
+import javax.json.bind.annotation.JsonbTransient;
 import javax.persistence.*;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlTransient;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -15,13 +15,13 @@ import java.util.List;
 @XmlRootElement
 @XmlAccessorType(XmlAccessType.FIELD)
 @Cacheable
-public class Catalog extends CatalogItem{
+public class Catalog extends CatalogItem {
 
     @ManyToMany(cascade = {CascadeType.PERSIST})
     @JoinTable(joinColumns = @JoinColumn(name = "catalogId"),
             inverseJoinColumns = @JoinColumn(name = "categoryId"))
     @OrderColumn(name="orderIdx")
-    @XmlTransient
+    @JsonbTransient
     private List<Category> rootCategories;
 
     @Transient
