@@ -3,11 +3,15 @@ package org.rembx.jeeshop.user;
 import com.querydsl.core.types.dsl.ComparableExpressionBase;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import io.quarkus.hibernate.orm.PersistenceUnit;
 import org.rembx.jeeshop.user.model.MailTemplate;
 import org.rembx.jeeshop.user.model.UserPersistenceUnit;
 
+import javax.annotation.Resource;
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.RequestScoped;
+import javax.inject.Inject;
 import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -17,11 +21,11 @@ import static org.rembx.jeeshop.user.model.QMailTemplate.mailTemplate;
 /**
  * Newsletter finder utility
  */
+@ApplicationScoped
 public class MailTemplateFinder {
 
     public final static String DEFAULT_LOCALE = "en_GB";
 
-    @PersistenceContext(unitName = UserPersistenceUnit.NAME)
     private EntityManager entityManager;
 
     private static final Map<String, ComparableExpressionBase<?>> sortProperties = new HashMap<String, ComparableExpressionBase<?>>() {{
@@ -32,10 +36,11 @@ public class MailTemplateFinder {
         put("updateDate", mailTemplate.updateDate);
     }};
 
-    public MailTemplateFinder() {
+    public static MailTemplateFinder getInstance(EntityManager entityManager) {
+        return new MailTemplateFinder(entityManager);
     }
 
-    public MailTemplateFinder(EntityManager entityManager) {
+    MailTemplateFinder( @PersistenceUnit(UserPersistenceUnit.NAME) EntityManager entityManager) {
         this.entityManager = entityManager;
     }
 
